Keep track of blobs that are part of multiple transactions (#7723)
* keep track of blobs that are part of multiple transactions
