+++ ethereum/core/src/test-support/java/org/hyperledger/besu/ethereum/core/BlobTestFixture.java
@@ -23,8 +23,6 @@ import org.hyperledger.besu.datatypes.KZGProof;
 import org.hyperledger.besu.datatypes.VersionedHash;
 import org.hyperledger.besu.evm.precompile.KZGPointEvalPrecompiledContract;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -35,6 +33,8 @@ import org.apache.tuweni.bytes.Bytes48;
 import org.bouncycastle.crypto.digests.SHA256Digest;
 
 public class BlobTestFixture {
+
+  private byte byteValue = 0x00;
 
   public BlobTestFixture() {
     try {
@@ -58,14 +58,8 @@ public class BlobTestFixture {
   ;
 
   public BlobTriplet createBlobTriplet() {
-    byte[] rawMaterial = {};
-    try (InputStream readme =
-        BlobTestFixture.class.getResourceAsStream(
-            "/org/hyperledger/besu/ethereum/core/encoding/BlobDataFixture.bin")) {
-      rawMaterial = readme.readAllBytes();
-    } catch (IOException e) {
-      fail("Failed to read blob file", e);
-    }
+    byte[] rawMaterial = new byte[131072];
+    rawMaterial[0] = byteValue++;
 
     Bytes48 commitment = Bytes48.wrap(CKZG4844JNI.blobToKzgCommitment(rawMaterial));
